feat: move fields to their own folder
